allauth adapter now thin. SocialAuth service
