+++ users/services/__init__.py
@@ -8,6 +8,7 @@ from .password_reset import (
     generate_password_reset_data,
     verify_password_reset_token,
 )
+from .social_auth import SocialAuthService
 from .utils import (
     get_user_from_uid,
     set_user_password,
@@ -22,4 +23,5 @@ __all__ = [
     'get_user_from_uid',
     'set_user_password',
     'verify_email_token',
+    'SocialAuthService',
 ]
